ci: use custom wait workflow
The existing workflow action does not handle multiple workflows on a repo.
Or at least, not without having to modify each repo workflow.

Add a WA using the code from kamilchodola/wait-for-workflow-action and
modify it allowing to specify a start time for the workflow.
Basically, we set the start time, then we create the workflow and then
we find the workflow which was deployed meanwhile, which should be 1.
If >1 we error out.

Also support cancelations, for easier testing.

Signed-off-by: Tiago Castro <tiagolobocastro@gmail.com>